backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
supabase==2.5.3
openai==1.35.3
librosa==0.10.2
spacy==3.7.5
textstat==0.7.3
boto3==1.34.141
httpx==0.27.0
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1
